requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
moviepy==1.0.3
pillow==10.0.0